requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21

# Database Drivers
psycopg2-binary==2.9.7  # PostgreSQL
PyMySQL==1.1.0          # MySQL (optional)

# Authentication & Security
PyJWT==2.8.1
bcrypt==4.0.1
python-dotenv==1.0.0

# Data Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Caching & Sessions
redis==5.0.0
Flask-Caching==2.0.2

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# API Documentation
Flask-RESTX==0.5.1

# Email
Flask-Mail==0.9.1

# Task Queue (Optional - for workers)
celery==5.3.2
celery-beat==2.5.0

# Payment Gateway Integration
stripe==5.16.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Monitoring & Logging
python-json-logger==2.0.7

# Code Quality
black==23.10.0
flake8==6.1.0
pylint==3.0.2
isort==5.12.0

# Development Tools
flask-cors==4.0.0  # CORS support
requests==2.31.0   # HTTP client for testing
